+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add generic links field to experiment plan. [#458](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/458)
+
 ### Changed
 
 - Change trailer and combine quick form log categories. [#438](https://github.com/Rothamsted-Ecoinformatics/farm_rothamsted/issues/438)